Refactored the code to improve reusability by creating Auth buttons using Button.nav.jsx and placing it into the organisms folder
