+++ src/components/organisms/Auth/LogoutButton.jsx
@@ -1,0 +1,18 @@
+import { useUser } from "../../../contexts/UserContext";
+import NavButton from "../../atoms/Button.nav";
+
+const LogoutButton = () => {
+  const { logout } = useUser();
+
+  const handleLogout = () => {
+    logout();
+  };
+
+  return (
+    <div>
+      <NavButton onClick={handleLogout}>Logout</NavButton>
+    </div>
+  );
+};
+
+export default LogoutButton;
